Add s3 service
Add s3-compatible service to dev/test stack.
Remove local file storage and always store
uploads in s3 or minio.

Change-type: major
