+++ apps/haproxy/haproxy.cfg
@@ -32,6 +32,9 @@ use_backend grafana_backend if host_grafana_backend
 acl host_prometheus_backend hdr_dom(host) -i prometheus.ly.fish.local
 use_backend prometheus_backend if host_prometheus_backend
 
+acl host_s3_backend hdr_dom(host) -i s3.ly.fish.local
+use_backend s3_backend if host_s3_backend
+
 frontend tcp_5432_in
     mode tcp
     bind *:5432
@@ -53,6 +56,11 @@ frontend tcp_9090_in
     mode tcp
     bind *:9090
     default_backend prometheus_backend
+
+frontend tcp_43697_in
+    mode tcp
+    bind *:43697
+    default_backend s3_backend_43697
 
 backend ui_backend
     mode http
@@ -81,6 +89,18 @@ backend registry_backend
     option forwardfor
     balance roundrobin
     server registry registry:80 resolvers docker-bridge-resolver resolve-prefer ipv4 check port 80
+
+backend s3_backend
+    mode http
+    option forwardfor
+    balance roundrobin
+    server s3 s3:80 resolvers docker-bridge-resolver resolve-prefer ipv4 check port 80
+
+backend s3_backend_43697
+    mode http
+    option forwardfor
+    balance roundrobin
+    server s3_43697 s3:43697 resolvers docker-bridge-resolver resolve-prefer ipv4 check port 43697
 
 backend grafana_backend
     mode http
