+++ package.json
@@ -34,7 +34,7 @@
     "compose:build": "docker-compose build",
     "compose:up": "docker-compose up",
     "compose:test": "docker-compose -f docker-compose.yml -f docker-compose.test.yml up",
-    "compose:database": "docker-compose up postgres redis haproxy",
+    "compose:data": "docker-compose up postgres redis s3 haproxy",
     "push": "./scripts/push.sh",
     "push:lib": "./scripts/push-lib.js",
     "build:server": "(cd apps/server && npm run build)",
@@ -77,7 +77,7 @@
   "devDependencies": {
     "@balena/jellyfish-client-sdk": "^10.3.3",
     "@balena/jellyfish-config": "^2.0.6",
-    "@balena/jellyfish-environment": "^9.1.19",
+    "@balena/jellyfish-environment": "^10.0.0",
     "@balena/jellyfish-plugin-balena-api": "^2.0.104",
     "@balena/jellyfish-plugin-channels": "^2.0.89",
     "@balena/jellyfish-plugin-default": "^24.0.74",
